+++ web/js/example.js
@@ -1,23 +1,31 @@
-
-// var toto = WEE.Ajax.call('assets/shaders/shader1.txt')
-//                    .then(function (val) { console.log(val); })
-//                    .catch(function (err) { console.log('error', err.message); });
-
-// var exampleDiv = Wee.Dom.findById('exampleDiv');
-// Wee.Dom.addHTMLElement(exampleDiv, 'p', {content: 'Ce paragraphe a été ajouté dynamiquement.', class: 'text-primary'});
-
-// var request = Aias.HTTP.get("assets/shaders/shader1.txt")
-//                        .then(function (val) { console.log(val); })
-//                        .catch(function (err) { console.log('error', err.message); });
-
 var data = {
-  firstname:'Galileo',
-  lastname:'Galilei',
-  born:1564,
-  died:1642
+  firstname: "Galileo",
+  lastname: "Galilei",
+  born: 1564,
+  died: 1642
 };
 
-var request = Aias.HTTP.post("http://httpbin.org/post", data)
-                       .then(function (response) { console.log(response); })
-                       .catch(function (err) { console.log('error', err.message); });
+var request = Aias.HTTP.GET("http://httpbin.org/get", "json")
+  .then(function(response) {
+    console.log(response);
+  })
+  .catch(function(err) {
+    console.log("error", err.message);
+  });
 
+var request = Aias.HTTP.POST("http://httpbin.org/post", "json", data)
+  .then(function(response) {
+    console.log(response);
+  })
+  .catch(function(err) {
+    console.log("error", err.message);
+  });
+
+Aias.HTTP.setEventType("observable");
+
+var request = Aias.HTTP.GET("http://httpbin.org/get", "json").subscribe(
+  response => {
+    console.log(response);
+  },
+  err => console.log("HTTP Error", err)
+);